move test to its own binary
